Better separate responsibilities
Before, the searchTerm was used for two cases: updating the input
field’s state and activating the side-effect for fetching data.

Now it’s only used for the former. A second state called url got
introduced for triggering the side-effect for fetching data which
only happens when a user clicks the confirmation button.