Improve last searches functionallity in a TDD approach
First, I added a test that fails proving that we have duplicated
buttons for duplicated searched terms.

Then I implemented what was necessary for the test to pass.

I've also added a test to make sure that the existing
functionallity of only showing the last 5 searched terms as button
still works.